+++ test-app/resources/perms/query.js
@@ -1,0 +1,3 @@
+if(query.$PREVENT_ALL) {
+    prevent('*');
+}