+++ test-app/resources/perms/validate.js
@@ -1,0 +1,5 @@
+if(this.title === '$ALLOW') {
+    allow('*');
+} else {
+    prevent('*');
+}