+++ HISTORY.md
@@ -7,5 +7,6 @@
 ### New Features
 
 - resources now support custom events by default
+- new collection permission api
 
 ### Major Bugfixes